Fixed audio suspension issue
Turns out the issue wasn't wuth the Nvidia settings, but rather with pipewire suspending audio sinks. I've corrected the fix_nvidia_audio.sh script so it actually runs and provides outputs on what it's doing. However, the real solution to the audio issue I was facing was to change the pipewire config to disable audio sink suspension. This was easiest to achieve via wireplumber.